requirements.txt
## The following requirements were added by pip freeze:
customtkinter==5.2.2
tksvg==0.7.4
yarg==0.1.10
mediapipe==0.10.33
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
pillow==12.2.0
pycaw==20251023
winrt-Windows.Devices.Radios==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1